+++ composer.json
@@ -29,8 +29,7 @@
         "symfony/filesystem": "^2.4|^3.0|^4.0",
         "symfony/finder": "^2.4|^3.0|^4.0",
         "symfony/options-resolver": "^2.6|^3.0|^4.0",
-        "symfony/process": "^2.1|^3.0|^4.0",
-        "symfony/web-profiler-bundle": "^4.0"
+        "symfony/process": "^2.1|^3.0|^4.0"
     },
     "require-dev": {
         "doctrine/dbal": "^2.4",
