+++ lib/Extension/ReportsExtension.php
@@ -15,7 +15,7 @@ namespace PhpBench\Extension;
 use PhpBench\DependencyInjection\Container;
 use PhpBench\DependencyInjection\ExtensionInterface;
 use PhpBench\Serializer\ElasticEncoder;
-use PhpBench\Storage\Driver\Reports\CurlTransport;
+use PhpBench\Storage\Driver\Reports\CurlTransmitter;
 use PhpBench\Storage\Driver\Reports\ReportsClient;
 use PhpBench\Storage\Driver\Reports\ReportsDriver;
 use PhpBench\Serializer\XmlEncoder;
@@ -31,20 +31,13 @@ class ReportsExtension implements ExtensionInterface
             return new ReportsDriver(
                 $container->get('storage.driver.reports.client'),
                 $container->get('storage.driver_registry'),
+                new XmlEncoder(),
                 $container->getParameter('storage.reports.inner_driver')
             );
         }, ['storage_driver' => ['name' => 'reports']]);
 
         $container->register('storage.driver.reports.client', function (Container $container) {
             return new ReportsClient(
-                $container->get('storage.driver.reports.transport'),
-                new XmlEncoder(),
-                $container->getParameter('storage.reports.store_iterations')
-            );
-        });
-
-        $container->register('storage.driver.reports.transport', function (Container $container) {
-            return new CurlTransport(
                 $container->getParameter('storage.reports.url'),
                 $container->getParameter('storage.reports.api_key')
             );
