+++ lib/Storage/Driver/Reports/ReportsDriver.php
@@ -17,6 +17,7 @@ use PhpBench\Model\Suite;
 use PhpBench\Model\SuiteCollection;
 use PhpBench\Registry\Registry;
 use PhpBench\Storage\DriverInterface;
+use PhpBench\Serializer\XmlEncoder;
 
 class ReportsDriver implements DriverInterface
 {
@@ -40,14 +41,21 @@ class ReportsDriver implements DriverInterface
      */
     private $innerStorageName;
 
+    /**
+     * @var XmlEncoder
+     */
+    private $xmlEncoder;
+
     public function __construct(
         ReportsClient $client,
         Registry $storageRegistry,
+        XmlEncoder $xmlEncoder,
         string $innerStorageName
     ) {
         $this->client = $client;
         $this->storageRegistry = $storageRegistry;
         $this->innerStorageName = $innerStorageName;
+        $this->xmlEncoder = $xmlEncoder;
     }
 
     /**
@@ -55,7 +63,9 @@ class ReportsDriver implements DriverInterface
      */
     public function store(SuiteCollection $collection)
     {
-        $response = $this->client->post($collection);
+        $suiteDocument = $this->xmlEncoder->encode($collection);
+        $response = $this->client->post('/import', $suiteDocument->dump());
+
         $this->innerDriver()->store($collection);
 
         return 'Report: ' . $response['suite_url'];
