--- lib/Storage/Driver/Reports/TransportInterface.php
@@ -1,18 +1,0 @@
-<?php
-
-/*
- * This file is part of the PHPBench package
- *
- * (c) Daniel Leech <daniel@dantleech.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- */
-
-namespace PhpBench\Storage\Driver\Reports;
-
-interface TransportInterface
-{
-    public function post(string $url, string $data): array;
-}
