+++ lib/Storage/DriverInterface.php
@@ -23,7 +23,11 @@ interface DriverInterface
     /**
      * Store the given SuiteCollection.
      *
+     * Optionally return a message which should be displayed
+     * by the CLI interface after successful storage.
+     *
      * @param SuiteCollection $collection
+     * @return string|null
      */
     public function store(SuiteCollection $collection);
